Move Feather metadata helpers into shared resources
Relocated the Feather metadata accessors from src/semantic/src/resources/feather-metadata.js to src/shared/src/resources/feather-metadata.js so the plugin no longer depends on semantic internals. Updated shared exports and plugin imports/tests to reference the new shared location.
