+++ src/plugin/src/ast-transforms/apply-feather-fixes.js
@@ -56,7 +56,7 @@ import {
     getFeatherDiagnosticById,
     getFeatherDiagnostics,
     getFeatherMetadata
-} from "gamemaker-language-semantic/resources/feather-metadata.js";
+} from "@prettier-plugin-gml/shared/resources/feather-metadata.js";
 import { loadReservedIdentifierNames } from "gamemaker-language-semantic/resources/reserved-identifiers.js";
 
 function walkAstNodes(root, visitor) {
